% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diabetes.R
\docType{data}
\name{diabetes}
\alias{diabetes}
\title{A Simulated Diabetes Dataset}
\format{
A data frame with 1000 rows and 4 variables:
\describe{
  \item{CLDL}{Change in low-density lipoprotein (LDL) cholesterol (mg/dL).}
  \item{GLUCOSE}{Change in fasting plasma glucose levels (mg/dL).}
  \item{TRIG}{Change in triglyceride levels (mg/dL).}
  \item{HBA1C}{Change in hemoglobin A1c levels (\%).}
}
}
\usage{
data("diabetes", package="MonotonicityTest")
}
\description{
This dataset contains simulated medical measurements for Diabetes and is
emulated after data from the Diabetes Prevention Program.
Each column represents change in a key metabolic indicators after two years for the placebo group
receiving no treatment.
}
\examples{
data("diabetes", package="MonotonicityTest")
names(diabetes)
}
\keyword{datasets}
