% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{monotonicity_test}
\alias{monotonicity_test}
\title{Do the monotonicity test}
\usage{
monotonicity_test(
  X,
  Y,
  bandwidth = bw.nrd(X) * (length(X)^-0.1),
  boot_num = 200,
  m = 10,
  ncores = 1
)
}
\arguments{
\item{X}{Vector of x values}

\item{Y}{Vector of Y values}

\item{bandwidth}{Kernel bandwidth}

\item{boot_num}{Number of bootstrap samples}

\item{m}{M parameter described in (paper)}

\item{ncores}{Number of cores to run on}
}
\value{
List with fields p, dist, stat
}
\description{
Does the monotonicity test as described in (Hall and Heckman)
}
